subminor change to required feat
Handle feat required for Resref construction. This is a non-functional change if Crafting.2da is well formed, in fact this entire code-block should eventually be removed.

Note that this will break Resref recipes if there is not a proper feat (other than Alertness) under SKILL. Before this change, malformed Resref recipes defaulted to a check for Craft Wondrous Items, even if the check should be for Craft Magical Arms and Armor. The default check is now for Alertness, which should be interpreted as Fix The Recipe.

NOTE: recompiled files have not been committed yet -- because this change is unlikely to affect anything (see above).