bypass (redundant) check for type-legal props
GetIsLegalItemProp() in GetRecipeMatches() [trigger: Imbue Item] ensured that any choices shown to the player were type-legal. But the check is also done in DoMagicCrafting() itself and shows a more accurate error message to the player, so let it happen there.

The code is simply commented out for now, it can be removed later if/when playtesting says things are satisfactory.