check ExcludedProp outside the valid/legal loop
not sure if things are better served by placing the ExcludedIP check before or after the loop. Right now it's after so that if a malformed (invalid) recipe shows up, that gets a hit first; however, the code could be considered both more efficient and readable if the ExcludedIP check were done before the loop.