define TCC_TYPE_BRACER
this change leaves the current behavior of Crafting.2da unchanged. That is,
TAGS="11" (or "11,1" etc.) still matches both gloves and bracer as before.
However, it's now possible to define a recipe for either gloves or
bracer that excludes the other wrist-type.

"36" - gloves exclusively
"78" - bracer exclusively
"11" - gloves or bracer (as before)

sidenote: The TCC-types for gloves and bracer are equivalent to baseitem
types as defined in BaseItems.2da. That is,

"36" is equivalent to "B36" and "78" is equivalent to "B78"